moved characters_init to character_loader.lua
